+++ packages/reactive/packages/decorator/unsubscribe/src/index.ts
@@ -1,0 +1,2 @@
+export { Unsubscribe } from './reactive-decorator.unsubscribe.func';
+export { unsubscribeOnDestroy } from './unsubscribe-on-destroy.func';
